+++ src/main.jsx
@@ -2,9 +2,12 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.jsx';
+import { TodosProvider } from './context/TodosContext';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <TodosProvider>
+      <App />
+    </TodosProvider>
   </StrictMode>
 );